+++ Gamecube/libgui/Button.h
@@ -80,7 +80,7 @@ private:
 	Image	*selectedFocusImage;
 	char** buttonText;
 	int buttonStyle, labelMode, labelScissor;
-	unsigned long StartTime;
+	u32 StartTime;
 	float x, y, width, height, fontSize;
 	GXColor	focusColor, inactiveColor, activeColor, selectedColor, labelColor;
 	ButtonFunc clickedFunc, returnFunc;
