+++ PeopsGXGPU/fps.h
@@ -33,7 +33,7 @@ void CheckFrameRate(void);
 void ReInitFrameCap(void);
 void SetAutoFrameCap(void);
 #ifndef _WINDOWS
-unsigned long timeGetTime();
+u32 timeGetTime();
 #endif
