+++ PeopsGXGPU/soft.h
@@ -43,16 +43,16 @@ void offsetPSX4(void);
 
 void FillSoftwareAreaTrans(short x0,short y0,short x1,short y1,unsigned short col);
 void FillSoftwareArea(short x0,short y0,short x1,short y1,unsigned short col);
-void drawPoly3G(long rgb1, long rgb2, long rgb3);
-void drawPoly4G(long rgb1, long rgb2, long rgb3, long rgb4);
-void drawPoly3F(long rgb);
-void drawPoly4F(long rgb);
+void drawPoly3G(s32 rgb1, s32 rgb2, s32 rgb3);
+void drawPoly4G(s32 rgb1, s32 rgb2, s32 rgb3, s32 rgb4);
+void drawPoly3F(s32 rgb);
+void drawPoly4F(s32 rgb);
 void drawPoly4FT(unsigned char * baseAddr);
 void drawPoly4GT(unsigned char * baseAddr);
 void drawPoly3FT(unsigned char * baseAddr);
 void drawPoly3GT(unsigned char * baseAddr);
-void DrawSoftwareSprite(unsigned char * baseAddr,short w,short h,long tx,long ty);
-void DrawSoftwareSpriteTWin(unsigned char * baseAddr,long w,long h);
-void DrawSoftwareSpriteMirror(unsigned char * baseAddr,long w,long h);
+void DrawSoftwareSprite(unsigned char * baseAddr,short w,short h,s32 tx,s32 ty);
+void DrawSoftwareSpriteTWin(unsigned char * baseAddr,s32 w,s32 h);
+void DrawSoftwareSpriteMirror(unsigned char * baseAddr,s32 w,s32 h);
 
 #endif // _GPU_SOFT_H_
